+++ blue-sol-backend/src/main/java/com/solif/backend/domain/youtube/service/YouTubeService.java
@@ -1,20 +1,10 @@
 package com.solif.backend.domain.youtube.service;
 
-import com.google.api.client.googleapis.javanet.GoogleNetHttpTransport;
-import com.google.api.client.json.gson.GsonFactory;
-import com.google.api.services.youtube.YouTube;
-import com.google.api.services.youtube.model.SearchListResponse;
-import com.google.api.services.youtube.model.SearchResult;
 import com.solif.backend.domain.youtube.dto.VideoDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.cache.annotation.CacheEvict;
-import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
-import java.io.IOException;
-import java.security.GeneralSecurityException;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -26,70 +16,7 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class YouTubeService {
     
-    @Value("${youtube.api.key}")
-    private String apiKey;
-    
-    @Value("${youtube.api.channel-id}")
-    private String channelId;
-    
-    /**
-     * YouTube API 클라이언트 생성
-     */
-    private YouTube getYouTubeService() throws GeneralSecurityException, IOException {
-        return new YouTube.Builder(
-            GoogleNetHttpTransport.newTrustedTransport(),
-            GsonFactory.getDefaultInstance(),
-            null
-        )
-        .setApplicationName("blue-sol-backend")
-        .build();
-    }
-    
-    /**
-     * 채널의 전체 동영상 조회 (캐싱 적용 - 고정 키)
-     * 최대 50개를 가져와서 캐싱하고, getLatestVideos에서 필요한 만큼만 반환
-     */
-    @Cacheable(value = "youtubeVideos", key = "'all'")
-    private List<VideoDto> getAllVideos() {
-        try {
-            log.info("YouTube API 호출: 전체 동영상 조회 (maxResults=50)");
-            
-            YouTube youtube = getYouTubeService();
-            
-            // 검색 요청 생성
-            YouTube.Search.List search = youtube.search()
-                .list(java.util.Arrays.asList("id", "snippet"))
-                .setChannelId(channelId)
-                .setKey(apiKey)
-                .setMaxResults(50L)  // 최대 50개 가져오기
-                .setOrder("date")  // 최신순
-                .setType(java.util.Arrays.asList("video")); // 동영상만
-            
-            // API 호출
-            SearchListResponse response = search.execute();
-            List<SearchResult> searchResults = response.getItems();
-            if (searchResults == null) {
-                return List.of();
-            }
-            
-            // DTO 변환
-            List<VideoDto> videos = searchResults.stream()
-                .map(VideoDto::fromSearchResult)
-                .filter(video -> video != null)  // null 제외
-                .collect(Collectors.toList());
-            
-            log.info("YouTube API 응답: {} 개의 동영상 조회됨", videos.size());
-            
-            // 카테고리 분류 적용
-            videos.forEach(this::assignCategory);
-            
-            return videos;
-            
-        } catch (Exception e) {
-            log.error("YouTube API 호출 실패", e);
-            throw new RuntimeException("YouTube 동영상 조회 실패: " + e.getMessage(), e);
-        }
-    }
+    private final YouTubeApiCacheService cacheService;
     
     /**
      * 채널의 최신 동영상 조회 (캐시된 전체 목록에서 필요한 만큼만 반환)
@@ -98,7 +25,10 @@ public class YouTubeService {
      * @return 동영상 목록
      */
     public List<VideoDto> getLatestVideos(Integer maxResults) {
-        List<VideoDto> allVideos = getAllVideos();  // 캐시에서 가져옴 (key='all')
+        List<VideoDto> allVideos = cacheService.getAllVideos();  // 캐시 서비스에서 가져옴
+        
+        // 카테고리 분류 적용 (캐시된 데이터에는 아직 적용 안됨)
+        allVideos.forEach(this::assignCategory);
         
         // maxResults만큼만 자르기
         return allVideos.stream()
@@ -108,16 +38,26 @@ public class YouTubeService {
     
     /**
      * 카테고리별 동영상 필터링
+     * 전체 동영상 목록에서 카테고리 필터링 후 maxResults만큼 반환
      */
     public List<VideoDto> getVideosByCategory(String category, Integer maxResults) {
-        List<VideoDto> allVideos = getLatestVideos(maxResults);
+        // 전체 동영상 조회 (캐시에서 최대 50개)
+        List<VideoDto> allVideos = cacheService.getAllVideos();
         
+        // 카테고리 분류 적용
+        allVideos.forEach(this::assignCategory);
+        
+        // "전체" 카테고리면 바로 maxResults만큼 반환
         if (category == null || category.equals("전체")) {
-            return allVideos;
+            return allVideos.stream()
+                .limit(maxResults)
+                .collect(Collectors.toList());
         }
         
+        // 카테고리 필터링 먼저 → 그 다음 limit 적용
         return allVideos.stream()
             .filter(video -> category.equals(video.getCategory()))
+            .limit(maxResults)  // ← 필터링 후 limit!
             .collect(Collectors.toList());
     }
     
@@ -192,11 +132,11 @@ public class YouTubeService {
      * 캐시 수동 갱신용 메서드
      * 기존 캐시를 모두 지우고 새로 조회하여 캐시 갱신
      */
-    @CacheEvict(value = "youtubeVideos", allEntries = true)
     public void refreshCache() {
         log.info("YouTube 동영상 캐시 갱신 시작");
         // 캐시를 지운 후 즉시 새로 조회하여 캐시 채우기
-        getAllVideos();
+        cacheService.evictCache();
+        cacheService.getAllVideos();
         log.info("YouTube 동영상 캐시 갱신 완료");
     }
 }
